CDS Scalar Function & usage
A CDS scalar function enables you to accept, for example, one or more scalar input parameters and to return a scalar value in CDS entities that are executable in the assigned engine. You can then make this function usable within regular ABAP CDS syntax in order to make a procedural logic available. It is a user defined that returns a scala value and always consist of a function definition and engine specific implementation.

Steps: 
1- Define the CDS scalar function definition
2- Define the CDS scalar function implementation reference (a class with interface if_amdp_marker_hdb)
3- Define the AMDP Function implementation

Source: 
-> https://www.youtube.com/watch?v=B9LxQIbzc2A&ab_channel=SAPDevelopers
-> https://www.youtube.com/watch?v=s8i44J4t7J0&ab_channel=SAPDevelopers 28:24
-> https://community.sap.com/t5/technology-blogs-by-sap/abap-cds-release-news-2308-cds-scalar-functions-implemented-by-amdp/ba-p/13561929
-> https://help.sap.com/docs/abap-cloud/abap-cds-tools-user-guide/cds-scalar-functions