CDS Propagated Buffer - for attention only
my current sap release not allowing it to use -> to take attention only 

Buffer propagation means that the joins defined by a CDS association are processed by the ABAP SQL engine on data stored in  the table buffer. Instead of using joins instantiated on the SAP HANA database, temporary joins are created on AS ABAP for processing the table buffer. This can lead to performance gains.

Source: 
-> https://help.sap.com/docs/abap-cloud/abap-data-models/cds-propagated-buffers
-> https://help.sap.com/doc/abapdocu_cp_index_htm/CLOUD/en-US/abencds_v2_view_buffering.html
-> https://www.youtube.com/watch?v=s8i44J4t7J0&ab_channel=SAPDevelopers 32:43